pyproject.toml
[project]
name = "fyp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt<5",
    "fastapi>=0.119.0",
    "groq>=0.31.1",
    "matplotlib>=3.10.8",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
